+++ src/LiveChartsCore/ColumnSeries.cs
@@ -74,7 +74,7 @@ namespace LiveChartsCore
             var previousSecondaryScale =
                 secondaryAxis.PreviousDataBounds is null ? null : new Scaler(drawLocation, drawMarginSize, secondaryAxis, true);
 
-            var uw = secondaryScale.ToPixels((float)secondaryAxis.UnitWidth) - secondaryScale.ToPixels(0f);
+            var uw = secondaryScale.ToPixels(secondaryAxis.UnitWidth) - secondaryScale.ToPixels(0f);
             var puw = previousSecondaryScale is null ? 0 : previousSecondaryScale.ToPixels((float)secondaryAxis.UnitWidth) - previousSecondaryScale.ToPixels(0f);
 
             uw -= (float)GroupPadding;
@@ -261,6 +261,7 @@ namespace LiveChartsCore
             CartesianChart<TDrawingContext> chart, IAxis<TDrawingContext> secondaryAxis, IAxis<TDrawingContext> primaryAxis)
         {
             var baseSeriesBounds = base.GetBounds(chart, secondaryAxis, primaryAxis);
+
             if (baseSeriesBounds.IsPrevious) return baseSeriesBounds;
             var baseBounds = baseSeriesBounds.Bounds;
