+++ src/LiveChartsCore/FinancialSeries.cs
@@ -111,7 +111,7 @@ namespace LiveChartsCore
             var previousSecondaryScale =
                 secondaryAxis.PreviousDataBounds is null ? null : new Scaler(drawLocation, drawMarginSize, secondaryAxis, true);
 
-            var uw = secondaryScale.ToPixels((float)secondaryAxis.UnitWidth) - secondaryScale.ToPixels(0f);
+            var uw = secondaryScale.ToPixels(secondaryAxis.UnitWidth) - secondaryScale.ToPixels(0f);
             var puw = previousSecondaryScale is null ? 0 : previousSecondaryScale.ToPixels((float)secondaryAxis.UnitWidth) - previousSecondaryScale.ToPixels(0f);
             var uwm = 0.5f * uw;
